Cleaned up code
- reorganized code to keep activity and auth0
  separate
