+++ app/src/main/java/com/foodmap/app/model/Auth0Manager.java
@@ -1,11 +1,6 @@
 package com.foodmap.app.model;
 
-import android.app.Activity;
 import android.content.Context;
-import android.content.Intent;
-import android.widget.Toast;
-
-import androidx.annotation.NonNull;
 
 import com.auth0.android.Auth0;
 import com.auth0.android.authentication.AuthenticationException;
@@ -18,64 +13,27 @@ public class Auth0Manager {
 
     private Auth0 account;
     private Context context;
-    private Activity activity;
-    private Class newClass;
 
-    public Auth0Manager(Context context, Activity activity, Class newClass){
+    public Auth0Manager(Context context){
         this.context = context;
-        this.activity = activity;
-        this.newClass = newClass;
-
         account = new Auth0(
                 context.getString(R.string.com_auth0_client_id),
                 context.getString(R.string.com_auth0_domain)
         );
     }
 
-    public void login(){
+    public void login(Callback<Credentials, AuthenticationException> callback){
         WebAuthProvider
                 .login(account)
                 .withScheme(context.getString(R.string.com_auth0_scheme))
-                .start(context, new Callback<Credentials, AuthenticationException>() {
-
-                    @Override
-                    public void onSuccess(Credentials credentials) {
-                        credentials.getUser().getId();
-                        System.out.println("login success");
-
-                        // open main activity
-                        Intent intent = new Intent(context, newClass);
-                        context.startActivity(intent);
-                        activity.finish();
-                    }
-
-                    @Override
-                    public void onFailure(@NonNull AuthenticationException e) {
-                        System.out.println("didnt work!");
-                        Toast.makeText(context,
-                                "Login failed: " + e.getMessage(), Toast.LENGTH_LONG).show();
-                    }
-                });
+                .start(context, callback);
     }
 
-    public void logout(){
+    public void logout(Callback<Void, AuthenticationException> callback){
         WebAuthProvider
                 .logout(account)
                 .withScheme(context.getString(R.string.com_auth0_scheme))
-                .start(context, new Callback<Void, AuthenticationException>() {
-
-                    @Override
-                    public void onSuccess(Void result) {
-                        // close main activity
-                        Intent intent = new Intent(context, newClass);
-                        context.startActivity(intent);
-                        activity.finish();
-                    }
-
-                    @Override
-                    public void onFailure(@NonNull AuthenticationException e) {
-                        System.out.println("Exception error: " + e.getMessage());
-                    }
-                });
+                .start(context, callback);
     }
+
 }
