+++ app/src/main/java/com/foodmap/app/ui/main/ProfileFragment.java
@@ -1,7 +1,9 @@
 package com.foodmap.app.ui.main;
 
+import android.content.Intent;
 import android.os.Bundle;
 
+import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 
 import android.view.LayoutInflater;
@@ -9,6 +11,8 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
 
+import com.auth0.android.authentication.AuthenticationException;
+import com.auth0.android.callback.Callback;
 import com.foodmap.app.R;
 import com.foodmap.app.model.Auth0Manager;
 import com.foodmap.app.ui.LoginActivity;
@@ -43,9 +47,26 @@ public class ProfileFragment extends Fragment {
         // Inflate the layout for this fragment
         View view = inflater.inflate(R.layout.fragment_profile, container, false);
 
-        auth0Manager = new Auth0Manager(getContext(), getActivity(), LoginActivity.class);
+        auth0Manager = new Auth0Manager(getContext());
         Button logoutBtn = view.findViewById(R.id.logoutBtn);
-        logoutBtn.setOnClickListener(v -> auth0Manager.logout());
+        logoutBtn.setOnClickListener(v -> {
+            auth0Manager.logout(new Callback<Void, AuthenticationException>() {
+
+                @Override
+                public void onSuccess(Void result) {
+
+                    // close main activity
+                    Intent intent = new Intent(getActivity(), LoginActivity.class);
+                    startActivity(intent);
+                    getActivity().finish();
+                }
+
+                @Override
+                public void onFailure(@NonNull AuthenticationException e) {
+                    System.out.println("Exception error: " + e.getMessage());
+                }
+            });
+        });
 
         return view;
     }
